+++ index.js
@@ -1,9 +1,10 @@
 import uuid from 'uuid/v4';
+import URITemplate from 'urijs/src/URITemplate';
+
 import URI from './src/uri';
 
-export {uuid};
 export changeFavicon from './src/changeFavicon';
 export {getBestLocale, getBrowserLocales} from './src/locales';
-export {URI};
+export {URI, URITemplate, uuid};
 export sanitizeFileName from './src/sanitizeFileName';
 export {convertSpringPropertyObject} from './src/convertSpringPropertyObject';